(lib) Fix lingering Path.walk callsites to use walk()
Use the custom impl walk() that only uses iterdir and is_file or is_dir

Fix lint errors
