+++ pathvein/lib.py
@@ -1,11 +1,10 @@
-from dataclasses import dataclass
 import logging
-import os
 from pathlib import Path
-from typing import Callable, Generator, Iterable, List, NamedTuple, Set, Tuple
+from typing import Callable, Iterable, List, NamedTuple, Set
 
 from upath import UPath
 
+from ._path_utils import walk
 from .pattern import FileStructurePattern
 
 logger = logging.getLogger(__name__)
@@ -30,36 +29,6 @@ class ShuffleResult(NamedTuple):
     source: Path
     destination: Path
     pattern: FileStructurePattern
-
-
-def walk(source: Path) -> Generator[Tuple[Path, List[str], List[str]], None, None]:
-    """
-    Recursively walk a directory path and return a list of directories and filesystem
-
-    Independent of os.walk or pathlib.Path.walk, this function just uses iterdir() to walk the directory tree
-
-    In this way it works for both pathlib.Path objects as well as third-party pathlib objects so long as they
-    implement iterdir(), is_dir, and is_file() methods.
-
-    This does not offer a sophisticated symlink following mechanism. If `type` is Path, this will short circuit
-    to os.walk which provides better symlink handling capability.
-    """
-    if type(source) is Path:
-        for dirpath, dirnames, filenames in os.walk(source):
-            yield Path(dirpath), dirnames, filenames
-    else:
-        dir_stack = []
-        dir_stack.append(source)
-        while dir_stack:
-            path = dir_stack.pop()
-            contents = list(path.iterdir())
-            filenames = [content.name for content in contents if content.is_file()]
-            dirs = [content for content in contents if content.is_dir()]
-            dirnames = [dir.name for dir in dirs]
-            yield path, dirnames, filenames
-            # Breadth-first traversal
-            logger.debug("dirs: %s", dirs)
-            dir_stack.extend(dirs)
 
 
 def scan(
